arch: riscv: 64 bit fixups
Fixed some issues in relation to 64 bit support:
tracking of natural integer size for CSR register values
Don't try and access PMPCFG 1 or 3 in 64 bit mode
as they do not exist and cause an exception.
Make protection area size match size of
is_user_mode as this is defined as a natural
registar sized integer.
make sure RISCV_PMP_CFG_NUM gets correct value if
CONFIG_PMP_SLOT is not multiple of 8 or 4.