arch: riscv: PMP_STACK_GUARD changes for arch_switch
This implements the changes for PMP_STACK_GUARD
from the arch_swap implementation in the
arch_switch code. When doing testing on this
I think the core_pmp.c code has some issues
which are not easy to fix but have probably
been addressed in the latest 3.x.x code base.
As the next phase of this development is to apply
what we have done here to the latest code base,
I'm leaving this as is and just pushing the
PMP_STACK_GUARD changes for reference.