arch: riscv: enable range of CPUs for SMP configuration
The CPU set for SMP is now more configurable to allow selection
of a contiguous group of CPUs.

We use a staged wake up sequence to allow monitoring the state
of each core as we bring them up and use wfi instructions for
the waiting cores to reduce impact on the shared bus when tightly
spinning. We can't rely on exiting wfi alone to indicate wake
up of a core as this can happen during debug sessions and so
we loop until the flag is set correctly.

Signed-off-by: Peter Mc Shane <peter.mcshane@microchip.com>