arch: riscv: add alias names for control status registers
RISC-V has different privilege levels such as Machine, Supervisor and
User. In order to bring up Zephyr in a privilege level other than
machine mode, we need acceess to the correct csr's for that privilege
level.

This PR introduces KConfig options to enable selecting Supervisor mode
and to enable booting from SBI such as OpenSBI.

As well as that, it will alias the CSR's that are used, based on the
privilege level Zephyr is to run in

Co-authored-by: Sven Ginka <s.ginka@sensry.de>
Signed-off-by: Conor Paxton <conor.paxton@microchip.com>