arch: riscv: use the cpu struct to get hartid
When running in supervisor mode, we still need to know what the hartid
is, in particular for the arch_proc_id() function. We also already know
the hartid of the boot hart from CONFIG_RV_BOOT_HART.

Also, when a HART boots into the Supervisor Execution
Environment (SEE), The higher priviliged firmware will put the hartid
of the hart that is about to be brought up in the a0 register, so there
is no need to write that into scratch, keeping scratch clean for
further use.

Expose hartid in cpu arch to enable it to be used in any mode, removing
the need to read mhartid directly

Signed-off-by: Conor Paxton <conor.paxton@microchip.com>