fixup: soc: mpfs: use supervisor external interrupts
This needs to be fixed and is solely used for testing. Turns out that
the plic driver grabs "only" the first IRQ num from the
interrupt-controller stanza of the device tree. Hart 0 (in our case)
only has Machine mode, so the plic pulls in the external interrupt num
for machine mode.

Coupled with that, the Kconfig symbol "2ND_LVL_INTR_00_OFFSET" defaults
to 11, trying to change the value in the device tree to 9 (supervisor
mode external interrupt) breaks compile

Signed-off-by: Conor Paxton <conor.paxton@microchip.com>