drivers: interrupt_controller: get context for supervisor mode
As Zephyr has always run in Macine mode, it is enough to return the
first context. Now that we want to run in Supervisor mode, we want to
return the correct context for it.

We do this by evaluating a symbol that gets set if the Zephyr
application is running in Supervisor mode. If so, return the correct
context

Signed-off-by: Conor Paxton <conor.paxton@microchip.com>